all pages

+++ application.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Catalog, SportItem
+from database_setup import Base, Catalog, SportItem, User
 from flask import session as login_session
 import random
 import string
@@ -35,25 +35,190 @@ def showLogin():
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
                     for x in xrange(32))
     login_session['state'] = state
-    # return "The current session state is %s" % login_session['state']
+    #return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+     # if user does not exist, create one
+    user_id = getUserID(data["email"])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    print getUserID(data["email"])
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 150px; height: 150px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    # Only disconnect a connected user.
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' %access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print url
+    print h
+    print result
+
+    if result['status'] == '200':
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
+# User Helper Functions
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
+# JSON APIs to view Catalog Information
+@app.route('/catalog/<int:catalog_id>/sport/JSON')
+def catalogSportJSON(catalog_id):
+    catalog = session.query(Catalog).filter_by(id=catalog_id).one()
+    items = session.query(SportItem).filter_by(
+        catalog_id=catalog_id).all()
+    return jsonify(SportItems=[i.serialize for i in items])
+
+
+@app.route('/catalog/<int:catalog_id>/sport/<int:sport_id>/JSON')
+def sportItemJSON(catalog_id, sport_id):
+    Sport_Item = session.query(SportItem).filter_by(id=sport_id).one()
+    return jsonify(Sport_Item=Sport_Item.serialize)
+
+
+@app.route('/catalog/JSON')
+def catalogsJSON():
+    catalogs = session.query(Catalog).all()
+    return jsonify(catalogs=[c.serialize for c in catalogs])
+
 
 # Show all catalog sports
 @app.route('/')
 @app.route('/catalog/')
 def showCatalogs():
     catalogs = session.query(Catalog).order_by(asc(Catalog.name))
+    items = session.query(SportItem).all()
+
     # if 'username' not in login_session:
     #     return render_template('publiccatalogs.html', catalogs=catalogs)
     # else:
-    return render_template('catalogs.html', catalogs=catalogs)
+    return render_template('catalogs.html', catalogs=catalogs, items=items)
 
 
 # Add a new sport
 @app.route('/catalog/new/', methods=['GET', 'POST'])
 def newCatalog():
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         newCatalog = Catalog(name=request.form['name'])
         session.add(newCatalog)
@@ -68,8 +233,8 @@ def newCatalog():
 
 @app.route('/catalog/<int:catalog_id>/edit/', methods=['GET', 'POST'])
 def editCatalog(catalog_id):
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     editedCatalog = session.query(
         Catalog).filter_by(id=catalog_id).one()
     if request.method == 'POST':
@@ -84,14 +249,16 @@ def editCatalog(catalog_id):
 # Delete a catalog
 @app.route('/catalog/<int:catalog_id>/delete/', methods=['GET', 'POST'])
 def deleteCatalog(catalog_id):
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+         return redirect('/login')
     catalogToDelete = session.query(
         Catalog).filter_by(id=catalog_id).one()
+
     if request.method == 'POST':
         session.delete(catalogToDelete)
-        flash('%s Successfully Deleted' % catalogToDelete.name)
         session.commit()
+        flash('%s Successfully Deleted' % catalogToDelete.name)
+
         return redirect(url_for('showCatalogs', catalog_id=catalog_id))
     else:
         return render_template('deleteCatalog.html', catalog=catalogToDelete)
@@ -105,7 +272,19 @@ def showItem(catalog_id):
     catalog = session.query(Catalog).filter_by(id=catalog_id).one()
     items = session.query(SportItem).filter_by(
         catalog_id=catalog_id).all()
-    return render_template('sport.html', items=items, catalog=catalog)
+    numOfItems = len(items)
+    catalogs = session.query(Catalog).order_by(asc(Catalog.name))
+
+    return render_template('sport.html', items=items, catalogs=catalogs, catalog=catalog, numOfItems=numOfItems)
+
+
+# Show a sport item description
+@app.route('/catalog/<int:catalog_id>/sport/<int:sport_id>/', methods=['GET', 'POST'])
+def showSportItem(catalog_id, sport_id):
+    displayedItem = session.query(SportItem).filter_by(id=sport_id).one()
+    catalog = session.query(Catalog).filter_by(id=catalog_id).one()
+
+    return render_template('sportItem.html', catalog=catalog, sport_id=sport_id, item=displayedItem)
 
 
 # Create a new sport item
@@ -122,11 +301,10 @@ def newSportItem(catalog_id):
         flash('New Sport %s Item Successfully Created' % (newItem.name))
         return redirect(url_for('showItem', catalog_id=catalog_id))
     else:
-        return render_template('newSportItem.html', catalog_id=catalog_id)
+        return render_template('newSportItem.html', catalog=catalog)
 
 
 # Edit a sport item
-
 @app.route('/catalog/<int:catalog_id>/sport/<int:sport_id>/edit', methods=['GET', 'POST'])
 def editSportItem(catalog_id, sport_id):
     if 'username' not in login_session:
